Improve classification functionality
